refactor: cleanup formattedTime
